Backend/requirements.txt
Flask==3.0.0
Flask_RESTful==0.3.10
flask_sqlalchemy==3.1.1
PyJWT==2.8.0
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
bcrypt==4.0.1
cloudinary==1.36.0
Flask-Cors==4.0.0
pipreqs==0.4.13
mysql
photos
pre-commit == 3.6.0
